Migrate upsertingUpdates to /api/
